Updated code including bug fixes and new features
The initial (parent) commit corresponds to the code submitted as part
of CQ 8780 (see https://dev.eclipse.org/ipzilla/show_bug.cgi?id=8780).
It was based on a snapshot of the previous code base as of 2014-07.

Since then the Capella code base has moved on with a lot of bug fixes
and new features.

Change-Id: I98604f9e8101238185d5e4e5cadea8462b48e3f1
Signed-off-by: Joao Barata <jean.barata@thalesgroup.com>