[255] Performance problem on Refresh of Duration in Exchange Scenario diagram
Fix performance problem caused by Acceleo3 default value

The default value of the attribute "Size Computation Expression" for the
style of an edge mapping is [eContents()->size()/]. This kind of
expression initialized the Acceleo3 interpreter that is not used in
Capella and this is costly.
This commit replaces all the default value of this attribute by "1".

Bug: 255
Change-Id: Ieaac1bf9da717f28c287edd350897c694f0d6de4
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>