[1055] Automagically generate .target file from setup model
Whenever setup is executed, a .target platform definition file for
capella dependencies is generated in the oomph folder. This .target file
can be used in a future tycho build. With this approach capella
dependencies only need to be maintained in the .setup model and desync
like seen before won't happen any more. After changing a dependency in
the setup model, just execute setup once and commit the updated .target
file.

Bug: 1055
Change-Id: Ifdb7fe1c661506b45373b70ba24e18dfca0941b8
Signed-off-by: Felix Dorner <felix.dorner@gmail.com>