[1055] Generate Capella.target into temp folder
The target file is first generated into temp. Then, only if "Capella
Developer" is installed, the target file is copied from the temporary
location into the releng plugin. Previously the target file was always
generated into the releng plugin - and that makes no sense if for
example only "Capella Addon Developer" is installed, since the releng
plugin does not exist in that case.

Change-Id: I1d5ccd018b61b99f44d508b5217adbe4fa59040c
Signed-off-by: Felix Dorner <felix.dorner@gmail.com>