[1322] Add tool to update connections between pairs of REC/RPL
This adds a new action "Update Connections" to the REC/RPL submenu. The
action is shown when a pair of RPL elements is selected. Executing the
action will compare the connections of corresponding REC elements with
those of the selected RPL elements and display differences in a dialog.
These differences can then be merged from the REC pair to the RPL pair
and vice versa.

Bug: 1322
Change-Id: I6f984e94d70e3ee6463491204c1ce64b91556e20
Signed-off-by: Felix Dorner <felix.dorner@gmail.com>