[1662]Fix save from the Capella Project Explorer view.
Instead of instantiating the CapellaSaveables directly,
CapellaSaveablesProvider relies on the Sirius extension point
org.eclipse.sirius.ui.uiSessionFactory to (ultimately) retrieve the
Saveables which define the save behavior to perform.
This ensures the consistency of the save behavior and also allows
extensions like Team For Capella to provide their own save
implementation.

Conflicts:
	core/plugins/org.polarsys.capella.core.platform.sirius.ui.navigator/src/org/polarsys/capella/core/platform/sirius/ui/navigator/view/CapellaCommonNavigator.java

Bug: 1662
Change-Id: Ic98fda24181917ba6273048a7f8ffc9f0e989560
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>
Signed-off-by: Philippe DUL <philippe.dul@thalesgroup.com>