[1673] Validate incoming/outgoing references during move
This adds two new validation rules I_37, I_38, one for live and one for
batch validation. In live mode, the rule detects which elements have
been moved during a transaction and verifies if the incoming/outgoing
references of these moved elements and their content
trees are still allowed. If a reference is no longer allowed, the
transaction rolls back and an error message is displayed to the user. In
batch mode, only outgoing references are verified. Batch mode is by
default disabled, live mode is by default enabled, yet only active when
the MoveElementsView is active.

Change-Id: I123b70232f4946bb697cf8c89439fb35cac9428f
Signed-off-by: Felix Dorner <felix.dorner@gmail.com>