[1745] Fix migration issue
* During Capella migration, a dedicated XMLHelper is used to load aird
resource. The consequence was that the ResourceDescriptor URI was
malformed (it only contained only the fragment URI without the resource
URI part) and then the Sirius migration participant, that migrates
DRepresentationDescriptor.repPath to reference the new
DRepresentation.uid attribute, failed. It led that
DRepresentationDescriptor.representation was always null.
* This commit adds the RepresentationsFileXMIHelper specificities to the
one used for Capella migration.

Bug: 1745
Change-Id: If7f82a7ba6a243328ca2c1c85e5157ec514a6f3d
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>