[1278] Better project creation API
This allows creation of capella project skeletons via builder pattern. I
did not refactor the current wizards to use the new builders, which
should be done in master, not in this patch version.

Still, the new builder is used in current and future tests and when
creating REC from library content.

Change-Id: Id201a285dce480509a43f85609f8abb1efc92daf
Signed-off-by: Felix Dorner <felix.dorner@gmail.com>