[527109] Disable WizardEditPolicy in Showing mode
This commit disables the Capella WizardEditPolicy when the Sirius
Showing mode is enabled. This allows to test this POC in Capella.
Without this fix, the wizard interferes with the double click policy
that is supposed to change the visibility state in this mode.

Bug: 527109
Change-Id: Ib2b67d9fddbb626be32627a1eaa872452e2bce77
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>
(cherry picked from commit 180c6386e5dd86850c64ff6729cd5783342e0432)