[1968] Allow to roll back the CapellaDeleteCommand if a delete hook detects an invalid state
Bug: 1968
This commit corresponds to a scenario available on T4C. In this
scenario the "hook.preDelete(getAllElementsToDelete())" returns a
"Status.CANCEL_STATUS" but this Cancel is not propagated. So the tool
continues to execute with invalid intermediate state and causes strange
behavior later. The idea of this commit is to really throw an
OperationCanceledException to really rollback the potential started
changes of all the command.

Conflicts:
	core/plugins/org.polarsys.capella.core.platform.sirius.ui.actions/src/org/polarsys/capella/core/platform/sirius/ui/commands/CapellaDeleteCommand.java

Conflicts:
	core/plugins/org.polarsys.capella.core.sirius.analysis/src/org/polarsys/capella/core/sirius/analysis/CapellaServices.java

Change-Id: I452c707ac380a0e73396d5e8161d9c1d9b47b65e
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>