[2041] Allow Multiple InstanceRoles with identical represented instances
This adds tools on ES, IS, OES scenarios to insert multiple
InstanceRoles for identical Component/Actor parts. When multiple
InstanceRoles with identical parts are present on a diagram, the name of
each InstanceRole is shown, in addition to the label of the represented
instance.

Transitions/refinements for such scenarios are not supported yet: The
user is asked to confirm or abort the launch of a multi-instancerole
scenario transition / refinement.

Change-Id: Ia9f72fbaec8f54ba4394f8cdfcfc3c68ca642871
Signed-off-by: Felix Dorner <felix.dorner@gmail.com>