[2104] Allow diagrams to reference model elements of interest
This adds a new property 'Elements of Interest' when
selecting the canvas of a diagram. Any model element can be
an element of interest for any diagram. Internally this
property is stored as a DAnnotation on the diagrams
RepresentationDescriptor, so it is possible to find elements
of interes without actually loading the diagram.

The semantic browser also shows all diagrams for which a
focused element is an element of interest.

Change-Id: Iea7f21c5adf8b79c01354ede039a4550d094b518
Signed-off-by: Felix Dorner <felix.dorner@gmail.com>