DRAFT [2249] For tests : Attempt to use a RecordingCommand implementation
It seems there is no need to use a specific command with the clone
listeners to react to undo/redo of the cloned repreentation command.

This draft is to clean the code and test the recording command approach.

The behavior of the undo/redo (and multiple undo/redo) before save must
be checked.

The behavior of the navigator refresh must be checked (see also activity
explorer and the other views displaying the rep descs.

The clone action was the only action using those listener whereas the
CreateRepresentations and Delete representations action has currently no
equivalent of this kind of mechanism

Conflicts:
	core/plugins/org.polarsys.capella.core.platform.sirius.ui.actions/src/org/polarsys/capella/core/platform/sirius/ui/commands/CapellaCloneDiagramCommand.java

Bug: 2249
Change-Id: Ic0169c83752ecbb2c667ad3cbc6ab342d937b1ba
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>