[2237] Use a SiriusCopier
The Clone command copies the annotations placed on the descriptors
before cloning the representations and their descriptors.

The viewpoint::DAnnotation is a concept from the Sirius core meta-model,
it inherits from viewpoint::IdentifiedElement and has the uid attribute
which is used as id and uriFragment. This id must be unique but the
EcoreUtil.Copier does not handle this (even if the attribute is declared
as id in the meta-model).

Note: this should be handled by Sirius.

Bug: 2237
Change-Id: I991712ca81941e2f31a939ca3baab7bf87d2f2d7
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>