[2271] "Sub/Super" -> "Generalized/Generalizing" + inheritance chain
Sub/Super categories for (Operational) Capabilities have been renamed to
Generalized/Generalizing Capabilities.

Also, when selecting a capability, the inheritance chain will be
displayed on the "Referenced/Referencing Elements" column in the
Semantic Browser.

EDIT: A bug was present inside AbstractCapabilityExt.java in method
getSubInheritanceHierarchyRecursively. The method would bring the sub
hierarchy as well as the super hierarchy. It has been fixed now to bring
only the sub hierarchy.

Bug: 2271
Change-Id: Ia3354f8f7f8400d747e591af74e2e7adf2e37bfe
Signed-off-by: Matei-Alexandru Stamate <matei-alexandru.stamate@thalesgroup.com>