[2287] Add QuickFix for EMF error concerning unresolved proxy
If a Capella element references another element in a library (class
generalization for example), and you delete it in the library, and you
validate your model referencing the library, an EMF error raises "The
feature X contains an unresolved proxy X".

Steps to reproduce:

    1) Create a library L;
    2) Create a class "Class Parent" inside L;
    3) Create a model M;
    4) Reference the library L;
    5) Create a class "Class Child" in M;
    6) Insert "Class Parent" and create a generalization from "Child" to
"Parent";
    7) Close the model M;
    8) Delete "Parent" in the library L;
    9) Open the model M;
    10) Validate the model M;

Actual Results:

An EMF error appears that concerns an unresolved proxy for the now
missing "Parent" class and no QuickFix is available for the error.

Expected Results:

A QuickFix is available for the error and, when applied, it deletes the
Generalization link that has its "super" missing.

Bug: 2287
Change-Id: I9f5a4f2be1ee70ec3676ca7384088a1620b45551
Signed-off-by: Matei-Alexandru Stamate <matei-alexandru.stamate@thalesgroup.com>