[2299] Validation rule + quickfix to detect duplicated Involvement links
Today, an Operational Capability can be involved twice in an Operational
Actor without any validation rule to detect.

Steps to reproduce:

1) Create an Operational Capabilities Blank (OCB) Diagram;
2) Create an Operational Capability;
3) Create an Operational Actor;
4) Add an Involvement Link between the Capability and the Actor;
5) From Capella Explorer, copy/paste the Involvement link of the
Operational Capability (disable the filter for
EntityOperationalCapabilityInvolvement if it  is active).
6) Validate the model.

Actual results:

The duplicate Involvement link is not detected by the validation.

Expected results:

The duplicate Involvement link is detected by the validation rules and a
quickfix is available to remove the duplicates.

EDIT: Also, fixed the bug where rule I_26 which detects duplicate Trace
elements was not working.

Change-Id: I0a99e8d45e30d7ccbcdc323f6e2f49968693dbfe
Signed-off-by: Matei-Alexandru Stamate <matei-alexandru.stamate@thalesgroup.com>