[2307] Validation rule + quick fix for EnumerationLiteral elements
If you drag&drop or cut/paste an EnumerationLiteral from an Enumeration
to another, the Type of the EnumerationLiteral will not change to match
the type of its new parent. This will introduce errors later in other
tools and it is an inconsistency that needs to be fixed.

Steps to reproduce:

1) Create a Class Diagram Blank (CDB);
2) Add 2 Enumeration Elements;
3) Add 2 EnumerationLiteral elements inside Enumeration1;
4) From Capella Explorer drag&drop EnumerationLiteral1 to Enumeration2;
5) From Capella Explorer cut/paste EnumerationLiteral2 to Enumeration2;
6) Add another EnumerationLiteral element inside Enumeration1;
7) From Properties change the Abstract Type to null for
EnumerationLiteral3;
8) Validate EnumerationLiteral 1, 2 and 3.

Actual Results:

There will be no errors/warnings present after the validation for
EnumerationLiteral1 and EnumerationLiteral2;

Rule DWF_D_35 will show a warning that EnumerationLiteral3 has no type.

Expected Results:

A validation rule will show a warning for EnumerationLiteral1 and
EnumerationLiteral2 saying that they should be typed by their parent
(Enumeration2). A QuickFix will be proposed which, when applied, will
change the Abstract Type property of EnumerationLiteral1 and
EnumerationLiteral2 to Enumeration2.

Rule DWF_D_35 will no longer be applied to EnumerationLiteral elements,
only to EnumerationReferences, because EnumerationReferences must have a
type but it is not mandatory for them to have the type of their parent.

Bug: 2307
Change-Id: I1737739a81debb722832bfd2fb90a9cb59dce065
Signed-off-by: Matei-Alexandru Stamate <matei-alexandru.stamate@thalesgroup.com>