[2389] Transfer view up and down buttons to select next/previous element
This adds up and down buttons to the transfer view stage viewer to
select the next/previous staged element or the next/previous element
with a backreference, whichever comes first. If the control key is
pressed when the action is executed, staged elements with a new parent
are skipped unless they have a backreference.

Bug: 2389
Change-Id: Ib92c34ede7ab0a3d57ee7c1a2d53437bb7127714
Signed-off-by: Felix Dorner <felix.dorner@gmail.com>