[Test] Fix tests due to composite FC enhancement
Do refresh all representations for the model EOLE_AF_UC after migration
so that the session will not be in dirty state after opening the diagram
(for instance [OPD] MainOperationalProcess - Operational Process
Description).

Otherwise, the session will be dirty after the first opening because of
some modification in terms of style.

- Fixed CheckDiagramDirtyStateOnOpeningTestCase

Change-Id: Iab9ea1bbf5df0a057b92ad0a1cf2f17501cfd951
Signed-off-by: cong-bang.do <docongbang1993@gmail.com>