[2477]  Unable to reference PV/PVG/EnumerationPropertyValue from library
Steps to reproduce:

1. Create a new Capella model;
2. Create a Library;
3. Reference the library in the model;
4. Create a PropertyValue, a PropertyValueGroup and an
EnumerationPropertyType inside the root SystemEngineering of the
library;
5. Try to assign the created elements to the Root System Function via
properties view;

Expected Results:

The propertyValue and propertyValueGroups should be available in the
dialog.

Actual Results:

The propertyValue and propertyValueGroup elements are not available in
the dialog.

Same for an EnumerationPropertyValue whose type you will try to set to
the created EnumerationPropertyType.

Bug: 2477
Change-Id: I034605ff3a84ddd9a267e8299b3caed4b70c8100
Signed-off-by: Matei-Alexandru Stamate <matei-alexandru.stamate@thalesgroup.com>