[draft] Cache by annotations
Main idea is to replace static methods helper by a static singleton
which is caching all methods annotated by @CacheResult.

It use dynamic proxy to create the class using the cache.
@see:
- ModelHelpers
- DataHelpers
- CacheInvovationHandler

It doesn't requires to caller to use getCache(class:method, param) way
to call a method. only the import is changed, and can be eclipse
content-assisted
with capella.epf file.

Change-Id: I024b994a7a74f322a251643956696a00b0ad3216
Signed-off-by: Philippe DUL <philippe.dul@thalesgroup.com>