[2524] Filter property sections via activity bindings
Property sections can now be filtered by activities,
bind it to an activity that matches 'sectionId'.

Tabs are hidden by default if there is no sections in it

Automatic refresh when activities change is not performed. The user has
to re-select the element to make tab changes visible in the view.

Bug: 2524
Change-Id: Idef35ec927c05553649ea2a16815329d8f184dea
Signed-off-by: Felix Dorner <felix.dorner@gmail.com>