[2608] SVG images in the Help section are sometimes not displayed
In the new Eclipse 2019-06 version the icons in the Help section have
been converted to svg (previously gif).

On some machines the internal Eclipse browser passes into compatibility
mode when opened, emulating a older version of the browser that may not
support SVG images.

Scenario:
- The user has IE 11 installed on his system.
- The internal browser of Eclipse goes into compatibility mode and
switches to IE7 for a unknown reason, leading to SVG images not being
able to be displayed.

The solution is to thus open the Help section in a external browser.

Bug: 2608
Change-Id: I58c75576e2982d173c57e2069c6b7620e059765b
Signed-off-by: Sandu Postaru <sandu.postaru@thalesgroup.com>