[DRAFT] Re-organize dependencies in manifest
The idea of this story is:

- Remove all "visibility:=reexport" as it makes dependencies implicit
and eventually developers does not pay attention where comes from the
dependency that they use.

- Remove unused dependencies for each plug-in: Overtime, certain
plug-ins are overloaded with unused dependencies whereas it should not.
For instance, while coding, developers add as many dependencies as they
want to experiment until found the wanted ones but they forget to remove
those unused before committing officially.

- If dependencies are well controlled, it would facilitate the re-factor
if needed and would avoid cycle dependency.

Change-Id: Ic97b81eac5347446dfbc62b61bdbf48ded71857d
Signed-off-by: cong-bang.do <docongbang1993@gmail.com>