[2692] Insert DataNotifier before the TransactionChangeRecorder.
The DataNotifier and the TransactionChangeRecorder are both adapters of
the ResourceSet which are then installed on each model element.

Their installation logic is different. The DataNotifier is a
EContentAdapter and is installed during xml resource load while
TransactionChangeRecorder is installed during notifyChanged() step. The
order on the model element comes from this installation order and might
be different from the ResourceSet.eAdapters() order.

On connected model, it is a bit different both adapters are installed
during notificationChanged() step in their order of registration on the
ResourceSet.

This commit install the DataNotifier before the
TransactionChangeRecorder in order to be sure to have the same relative
order in all situations.

Bug: https://bugs.polarsys.org/show_bug.cgi?id=2692
Change-Id: Ie35df2f89b368719ced0f2c7dabb5ef9290fac01
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>
Signed-off-by: Philippe DUL <philippe.dul@thalesgroup.com>