[553820] Override default values for jump links properties
This commit allows to override default values for jump links properties
and have by default a "tunnel" jump link enabled.

These preferences have not been overridden in
org.polarsys.capella.common.platform.sirius.customisation.SiriusPlugin
because they are not overridden as soon as Capella is launched. A
session must be opened to override default value. This is the same for
all preferences overridden in start of SiriusPlugin.

Bug: 553820
Change-Id: I8607afd585d42539158edaa4da5fa5259bfbbcaf
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>