[releng] Fix compilation problem for 2020-06 TP concerning javax.inject
Since platform 2019-09, there is a compilation problem in
ToolbarPopulator class. The import javax.inject.Inject can not be
resolved.

The reason is the change made by Eclipse bug 487676 [1] which removed
the 're-export' of the javax.inject dependency from
org.eclipse.core.runtime so that it now has to explicitly be added in
dependency. This is also documented in the Whats New for Eclipse 2019-09
(4.13) [2].

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=487676
[2] https://www.eclipse.org/eclipse/news/4.13/platform_isv.php

Change-Id: I814c4048460d23273163ec8a47e53320e12d3835
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>