[test] Avoid the installation of the test DAnalysisSelector
The org.polarsys.capella.test.diagram.common.ju plugin declares a
TestSessionManagerListener which will inject a test DAnalysisSelector in
the Sirius Session.
It is sometimes painful when working in a development context because it
changes the runtime behavior and breaks a lot of tests.

This commit allows to deactivate the TestSessionManagerListener using
the system property "capella.tests.disablecustomdanalysisselector".

Change-Id: I07290b600abfa0836c8c21ec4b2faa3be987d5b7
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>