[567909] Improve analysis selector to be more user friendly
There as been some improvement on the analysis selector to be more user
friendly. As the one provided in Capella was better that the one in
Sirius, it was made as a class (instead of a nested class) so it could
be reused properly.
Furthermore, the move representation actions have been modified. Instead
of having an action for each available aird file, there is now only one
action that displays the analysis selector.

Bug: 567909
Change-Id: I8398171709a09548a863fd8edd1d163addc2b3a6
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>
Signed-off-by: Philippe DUL <philippe.dul@thalesgroup.com>