#2379 Add icons and text for StateFragment
Rely on the icons and label of the StateFragment related object
(Function or State) to display in the ProjectExplorer.
Added an overlay on new StateFragmetn icons (when a related element is
set on the StateFragment).
Defaults to State icon/name if both state and function are provided.

Update default icon for StateFragment, used when none of the related
elements are set.

Fix minor code smell.

Change-Id: I228fd4ff1c19687928dc40411a373b975e1547e2
Signed-off-by: Arnaud Dieumegard <arnaud.dieumegard@obeo.fr>