#2406 Allow to specify expected pre-rename model name
WorkspaceImagePathChange is expected to be triggered during a Capella
project rename change but cannot be used to apply the refactoring on an
already renamed project.

Provide a new constructor with the "source" project name to be replaced
in the content of the description fields. Adapt the replace behavior
when using this constructor.

Change-Id: I0000000000000000000000000000000000000000
Signed-off-by: Arnaud Dieumegard <arnaud.dieumegard@obeo.fr>