#2410 Handle project name update in images paths for transition
Ensure that image paths in the model are updated accordingly to the
transition configuration (reference and target).

Here we integrate the use of an overriden version of the
SiriusImageHelper from DiffMerge The overriden version allows to specify
the resource used to retrieve the project name and may also be
deactivated.

The default Reference and Target Transition Scopes have been augmented
with code adaptation from SiriusImageHelper and initialize the
TransitionSiriusImageHelper according to their needs.
The ReScope (for REC/RPL management) and the ReferenceModelScope and
TargetModelScopes for topdown transition have been adapted to avoid the
application of the code adaptation as in these cases, no modification of
the paths shall be done.

See https://github.com/eclipse/org.eclipse.emf.diffmerge.core/issues/234
for the related DiffMerge modifications.

Change-Id: I0000000000000000000000000000000000000000
Signed-off-by: Arnaud Dieumegard <arnaud.dieumegard@obeo.fr>