#2449 Navigation towards related diagrams on double-click
Added new preference in Capella>Diagram pref page

This preference decides whether or not a double click on particular
elements shall perform navigation towards its related diagram, or if it
shall use the default behaviour (open properties wizard)

This preference is enabled by default

These "particular elements" are :
- FunctionalChain and FunctionalChainReference
- PhysicalPath and PhysicalPathReference
- InteractionUse
- Scenario

================================================================

Added navigation on double click in diagram view
Added an editpolicy performing the navigation on double click, in the
diagram view, on these elements.
The policy is always added, but the command it provides is enabled
according to the preference value.
This allows navigation to be enabled/disabled through the preference
page, without having to reload (close and reopen) the diagram, as the
edit policies are attached on diagram opening.

Modified the default policy so that it isn't enabled on these particular
elements when the other policy is enabled

================================================================

Added navigation on double click in navigator

Changed OpenPropertiesAction to OpenAction
so that it handles both 'open properties wizard' and 'navigate to
related diagrams' actions as there is no way to register two actions

The navigation to related diagrams is done if the preference is enabled
and if the selected element is in the list of "particular elements"

================================================================

Added navigation behaviour on double click in semantic browser

================================================================
Added dialogs for navigation behaviour on double-click

in navigator, semantic browser and diagram view

Whenever a "particular" element is double-clicked in the navigator, when
navigation on double click is activated :

-If the element has no existing representation
--If multiple representation are available, display a dialog that let
the user choose a new representation to create and let it name it (using
the default name whenever a representation is selected)
--If only one representation is available, display the "new
representation" dialog with a message "No existing representation, would
you like to create one ?"
(Added validation error when name is empty on these dialogs)

- IF the element has existing representation
--If there is only one, open it (behaviour already implemented)
--If there are several, display a dialog that let the user choose which
representation to open

================================================================

Updated Documentation

Added information about the "Navigate to related Diagrams on double
click" preference

Added information on new double click behaviour

================================================================

Updated RCPTT

All "specific" elements are tested in all views (Semantic Browser,
Navigator /Project Explorer, Diagram Editor)

All tested cases :
- Only 1 existing diagram
- Several existing diagrams
- No existing diagram - Only one diagram type available

The case with no existing diagram and multiple diagram types available
cannot be tested as none of the "specific" elements are eligible

Change-Id: Iddaa54e9fb68576f75d67165db634434d0882f85
Signed-off-by: Erwann Traisnel <erwann.traisnel@obeo.fr>