[#2449] Navigation towards related diagrams on double-click in diagram view
Added new preference in Capella>Diagram pref page
This preference decides whether or not a double click on particular
elements shall perform navigation towards its related diagram, or if it
shall use the default behaviour (open properties wizard)

This preference is enabled by default

These "particular elements" are :
- FunctionalChain and FunctionalChainReference
- PhysicalPath and PhysicalPathReference
- InteractionUse
- Scenario

Added an editpolicy performing the navigation on double click, in the
diagram view, on these elements.
The policy is always added, but the command it provides is enabled
according to the preference value.
This allows navigation to be enabled/disabled through the preference
page, without having to reload (close and reopen) the diagram, as the
edit policies are attached on diagram opening.

Modified the default policy so that it isn't enabled on these particular
elements when the other policy is enabled

Change-Id: I9166003b881328bcccad9185f79b8cfc336863c3
Signed-off-by: Erwann Traisnel <erwann.traisnel@obeo.fr>