#2514 Update library property tester and provider/manager APis
To bypass the limitation of library management for shared capella
project (add/remove reference to shared library project), a property
tester requires to be updated and an API added:
- CapellaModelIsLocalPropertyTester now uses the airdFile to check that
the session is opened and contains a Capella model
- ILibraryManager and ILibraryProvider methods getAvailableModels has
been doubled to give the session has a parameter. This allows
CDOLibraryProvider to use the session to check for other models on the
CDO repository it is connected to.
- Addition of a method isLocalProject in AbstractCapellaProvider so
CapellaWorkspaceProvider can provide local libraries only to local
Capella project.
- remove cache of models from LibraryManager as switching from a local
project to a shared project caused the list of available libraries to be
wrong. Also the cache showed a wrong list when all libraries reference
where removed.

Change-Id: If79cedfe0128cc960f3da679706c5be85557c1e2
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>