#2913 Make association source and target reliable in CDB
This commit updates the CDB association edge mapping in the VSM
(source/target mappings and styles) to make association source and
target reliable in CDB, to always have the same source and target, and
to prevent a change in property from changing the source or the target.

This commit also adds a migration to reverse the association edge in the
wrong direction in existing projects.

This commit also adapts the tests
ReconnectRelationshipGroup.ReconnectRelationshipGroupSA,
ReconnectRelationshipGroup.ReconnectRelationshipGroupOA in
org.polarsys.capella.test.diagram.tools.ju.cdb to these changes.

Signed-off-by: Séraphin Costa <seraphin.costa@obeosoft.com>